Convert To TypeScript
