+++ backend/package.json
@@ -20,6 +20,8 @@
     "ts-node-dev": "^1.0.0-pre.44"
   },
   "dependencies": {
+    "@types/cors": "^2.8.6",
+    "@types/express": "^4.17.6",
     "cors": "^2.8.5",
     "express": "^4.17.1",
     "sqlite3": "^4.2.0",
